Refactor feedback module to normalized 3-table design with anonymous ratings

+++ phantask-backend/src/main/java/com/phantask/feedback/dto/FeedbackReportDto.java
@@ -1,24 +1,42 @@
 package com.phantask.feedback.dto;
 
-import java.time.LocalDate;
-import java.time.LocalDateTime;
 import java.util.Map;
-
-import com.phantask.task.entity.TaskStatus;
-
 import lombok.Data;
-import lombok.Getter;
-import lombok.Setter;
 
+/**
+ * DTO returned to ADMIN when viewing feedback report.
+ *
+ * This DTO contains only aggregated data.
+ * Individual user ratings or identities are never exposed.
+ */
 @Data
 public class FeedbackReportDto {
 
-	private Map<String, Integer> averagePerQuestion; // rounded (1–5)
-    private int overallAverage;                      // rounded (1–10)
-    private int totalSubmissions;
-    
-    
-	
-    /* getters & setters */
-    
+    /**
+     * Average rating per question.
+     * Values are rounded and range from 1 to 5.
+     *
+     * Example:
+     * {
+     *   "Cleanliness": 4,
+     *   "Food Quality": 5
+     * }
+     */
+    private Map<String, Integer> averagePerQuestion;
+
+    /**
+     * Overall average rating scaled to 10.
+     * Calculation:
+     *  - Average of per-question averages
+     *  - Multiply by 2
+     *  - Round off
+     *
+     * Example: 8
+     */
+    private int overallAverage;
+
+    /**
+     * Total number of users who submitted this feedback.
+     */
+    private long totalSubmissions;
 }
